+++ CHANGELOG.md
@@ -32,6 +32,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added `ModuleManifest.common.v5.Tests.ps1`
   - Added `ModuleScriptFiles.common.v5.Tests.ps1`
     - Contain the converted Pester 4 tests from `Psm1Parsing.common.v4.Tests.ps1`.
+  - Added `PSSAResource.common.v5.Tests.ps1`
+    - Any test that is excluded by using Pester `ExcludeTag` under the key
+      `DscTest:` will now be silently excluded due to how Pester does _Discovery_.
 
 ### Changed